local_test/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
